+++ src/test/java/org/jabref/logic/citationstyle/CitationStyleGeneratorTest.java
@@ -1,17 +1,19 @@
 package org.jabref.logic.citationstyle;
 
+import java.util.List;
+
 import org.jabref.logic.l10n.Localization;
 import org.jabref.logic.util.TestEntry;
 import org.jabref.model.database.BibDatabase;
 import org.jabref.model.entry.BibEntry;
 import org.jabref.model.entry.field.StandardField;
+import org.jabref.model.entry.types.StandardEntryType;
 
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-@Disabled("For some reason, instead of vol and pp we get null. No idea about the origin of this problem.")
 class CitationStyleGeneratorTest {
 
     @Test
@@ -21,7 +23,7 @@ class CitationStyleGeneratorTest {
 
         // if the default citation style changes this has to be modified
         String expected = "  <div class=\"csl-entry\">\n" +
-                "    <div class=\"csl-left-margin\">[1]</div><div class=\"csl-right-inline\">F. Last and J. Doe, .</div>\n" +
+                "    <div class=\"csl-left-margin\">[1]</div><div class=\"csl-right-inline\">F. Last and J. Doe, </div>\n" +
                 "  </div>\n";
         String citation = CitationStyleGenerator.generateCitation(entry, CitationStyle.getDefault());
         assertEquals(expected, citation);
@@ -34,7 +36,7 @@ class CitationStyleGeneratorTest {
 
         // if the default citation style changes this has to be modified
         String expected = "  <div class=\"csl-entry\">\n" +
-                "    <div class=\"csl-left-margin\">[1]</div><div class=\"csl-right-inline\">F. Last and J. Doe, .</div>\n" +
+                "    <div class=\"csl-left-margin\">[1]</div><div class=\"csl-right-inline\">F. Last and J. Doe, </div>\n" +
                 "  </div>\n";
         String citation = CitationStyleGenerator.generateCitation(entry, CitationStyle.getDefault());
         assertEquals(expected, citation);
@@ -47,18 +49,9 @@ class CitationStyleGeneratorTest {
         assertEquals(expected, citation);
     }
 
-    @Test
-    void testAsciiDocFormat() {
-        String expectedCitation = "[1] B. Smith, B. Jones, and J. Williams, ``Title of the test entry,'' __BibTeX Journal__, vol. 34, no. 3, pp. 45–67, Jul. 2016.\n";
-        BibEntry entry = TestEntry.getTestEntry();
-        String style = CitationStyle.getDefault().getSource();
-        CitationStyleOutputFormat format = CitationStyleOutputFormat.ASCII_DOC;
-
-        String actualCitation = CitationStyleGenerator.generateCitation(entry, style, format, new BibDatabase());
-        assertEquals(expectedCitation, actualCitation);
-    }
 
     @Test
+    @Disabled("Currently citeproc does not handler number field correctly https://github.com/JabRef/jabref/issues/8372")
     void testHtmlFormat() {
         String expectedCitation = "  <div class=\"csl-entry\">\n" +
                 "    <div class=\"csl-left-margin\">[1]</div><div class=\"csl-right-inline\">B. Smith, B. Jones, and J. Williams, “Title of the test entry,” <i>BibTeX Journal</i>, vol. 34, no. 3, pp. 45–67, Jul. 2016.</div>\n" +
@@ -71,18 +64,9 @@ class CitationStyleGeneratorTest {
         assertEquals(expectedCitation, actualCitation);
     }
 
-    @Test
-    void testRtfFormat() {
-        String expectedCitation = "[1]\\tab B. Smith, B. Jones, and J. Williams, \\uc0\\u8220{}Title of the test entry,\\uc0\\u8221{} {\\i{}BibTeX Journal}, vol. 34, no. 3, pp. 45\\uc0\\u8211{}67, Jul. 2016.\r\n";
-        BibEntry entry = TestEntry.getTestEntry();
-        String style = CitationStyle.getDefault().getSource();
-        CitationStyleOutputFormat format = CitationStyleOutputFormat.RTF;
-
-        String actualCitation = CitationStyleGenerator.generateCitation(entry, style, format, new BibDatabase());
-        assertEquals(expectedCitation, actualCitation);
-    }
 
     @Test
+    @Disabled("Currently citeproc does not handle number field correctly https://github.com/JabRef/jabref/issues/8372")
     void testTextFormat() {
         String expectedCitation = "[1]B. Smith, B. Jones, and J. Williams, “Title of the test entry,” BibTeX Journal, vol. 34, no. 3, pp. 45–67, Jul. 2016.\n";
         BibEntry entry = TestEntry.getTestEntry();
@@ -94,45 +78,21 @@ class CitationStyleGeneratorTest {
     }
 
     @Test
-    void testXslFoFormat() {
-        String expectedCitation = "<fo:block id=\"Smith2016\">\n" +
-                "  <fo:table table-layout=\"fixed\" width=\"100%\">\n" +
-                "    <fo:table-column column-number=\"1\" column-width=\"2.5em\"/>\n" +
-                "    <fo:table-column column-number=\"2\" column-width=\"proportional-column-width(1)\"/>\n" +
-                "    <fo:table-body>\n" +
-                "      <fo:table-row>\n" +
-                "        <fo:table-cell>\n" +
-                "          <fo:block>[1]</fo:block>\n" +
-                "        </fo:table-cell>\n" +
-                "        <fo:table-cell>\n" +
-                "          <fo:block>B. Smith, B. Jones, and J. Williams, “Title of the test entry,” <fo:inline font-style=\"italic\">BibTeX Journal</fo:inline>, vol. 34, no. 3, pp. 45–67, Jul. 2016.</fo:block>\n" +
-                "        </fo:table-cell>\n" +
-                "      </fo:table-row>\n" +
-                "    </fo:table-body>\n" +
-                "  </fo:table>\n" +
-                "</fo:block>\n";
-        BibEntry entry = TestEntry.getTestEntry();
-        String style = CitationStyle.getDefault().getSource();
-        CitationStyleOutputFormat format = CitationStyleOutputFormat.XSL_FO;
-
-        String actualCitation = CitationStyleGenerator.generateCitation(entry, style, format, new BibDatabase());
-        assertEquals(expectedCitation, actualCitation);
-    }
-
-    @Test
     void testHandleDiacritics() {
         BibEntry entry = new BibEntry();
-        entry.setField(StandardField.AUTHOR, "L{\"a}st, First and Doe, Jane");
+        // We need to escape the backslash as well, because the slash is part of the LaTeX expression
+        entry.setField(StandardField.AUTHOR, "L{\\\"a}st, First and Doe, Jane");
         // if the default citation style changes this has to be modified.
         // in this case ä was added to check if it is formatted appropriately
         String expected = "  <div class=\"csl-entry\">\n" +
-                "    <div class=\"csl-left-margin\">[1]</div><div class=\"csl-right-inline\">F. Läst and J. Doe, .</div>\n" +
+                "    <div class=\"csl-left-margin\">[1]</div><div class=\"csl-right-inline\">F. L&auml;st and J. Doe, </div>\n" +
                 "  </div>\n";
         String citation = CitationStyleGenerator.generateCitation(entry, CitationStyle.getDefault());
         assertEquals(expected, citation);
     }
 
     @Test
+    @Disabled("Currently citeproc does not handler number field correctly")
     void testHandleAmpersand() {
         String expectedCitation = "[1]B. Smith, B. Jones, and J. Williams, “&TitleTest&” BibTeX Journal, vol. 34, no. 3, pp. 45–67, Jul. 2016.\n";
         BibEntry entry = TestEntry.getTestEntry();
@@ -143,4 +103,32 @@ class CitationStyleGeneratorTest {
         String actualCitation = CitationStyleGenerator.generateCitation(entry, style, format, new BibDatabase());
         assertEquals(expectedCitation, actualCitation);
     }
+
+    @Test
+    void testHandleCrossRefFields() {
+
+        BibEntry firstEntry = new BibEntry(StandardEntryType.InCollection)
+            .withCitationKey("smit2021")
+            .withField(StandardField.AUTHOR, "Smith, Bob")
+            .withField(StandardField.TITLE, "An article")
+            .withField(StandardField.PAGES, "1-10")
+            .withField(StandardField.CROSSREF, "jone2021");
+
+        BibEntry secondEntry = new BibEntry(StandardEntryType.Book)
+            .withCitationKey("jone2021")
+            .withField(StandardField.EDITOR, "Jones, John")
+            .withField(StandardField.PUBLISHER, "Great Publisher")
+            .withField(StandardField.TITLE, "A book")
+            .withField(StandardField.YEAR, "2021")
+            .withField(StandardField.ADDRESS, "Somewhere");
+
+        String expectedCitation = "[1]B. Smith, “An article,” J. Jones, Ed. Somewhere: Great Publisher, 2021, pp. 1–10.\n";
+        BibDatabase bibDatabase = new BibDatabase(List.of(firstEntry, secondEntry));
+        String style = CitationStyle.getDefault().getSource();
+
+        String actualCitation = CitationStyleGenerator.generateCitation(firstEntry, style, CitationStyleOutputFormat.TEXT, bibDatabase);
+        assertEquals(expectedCitation, actualCitation);
+
+    }
 }
+